+++ pkg/db/client.go
@@ -152,7 +152,8 @@ func (client *Client) SavePageViews(ctx context.Context, pageViews []model.PageV
 	query, err := tx.PrepareContext(ctx, `INSERT INTO "page_view" (client_id, visitor_id, session_id, time, duration_seconds,
 		path, title, language, country_code, city, referrer, referrer_name, referrer_icon, os, os_version,
 		browser, browser_version, desktop, mobile, screen_class,
-		utm_source, utm_medium, utm_campaign, utm_content, utm_term) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`)
+		utm_source, utm_medium, utm_campaign, utm_content, utm_term,
+		tag_keys, tag_values) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)`)
 
 	if err != nil {
 		return err
@@ -183,7 +184,9 @@ func (client *Client) SavePageViews(ctx context.Context, pageViews []model.PageV
 			pageView.UTMMedium,
 			pageView.UTMCampaign,
 			pageView.UTMContent,
-			pageView.UTMTerm)
+			pageView.UTMTerm,
+			pageView.TagKeys,
+			pageView.TagValues)
 
 		if err != nil {
 			if e := tx.Rollback(); e != nil {